refactored ready for tests
