+++ old_code/run_economy.py
@@ -10,9 +10,6 @@ from tqdm import tqdm  ### package for progress bars
 os.chdir(".")
 import matplotlib.pyplot as plt
 from economy_class_wealth import Model1
-from inequality_metrics import find_wealth_groups
-from enkf_yo import EnsembleKalmanFilter
-
 
 #%%
 ### LOAD empirical monthly wealth Data
