+++ src/enkf_yo.py
@@ -5,6 +5,7 @@ Created on Tue Apr 25 13:47:43 2023
 @author: earyo
 """
 # Imports
+import os
 import warnings as warns
 import numpy as np
 import pandas as pd
@@ -90,7 +91,7 @@ class EnsembleKalmanFilter:
         ### load observation data from desired start year (y)
         ### LOAD empirical monthly wealth Data sorted by group
         ### for state vector check
-        with open('./data/wealth_data_for_import2.csv') as f2:
+        with open(os.path.join('data', 'wealth_data_for_import2.csv')) as f2:
             self.data = pd.read_csv(f2, encoding = 'unicode_escape')        
         y = model_params["start_year"]
         self.idx_begin = min((self.data[self.data["year"]==y].index.values))
