+++ src/enkf_yo2.py
@@ -92,7 +92,7 @@ class EnsembleKalmanFilter:
         ### load observation data from desired start year (y)
         ### LOAD empirical monthly wealth Data sorted by group
         ### for state vector check
-        with open('./data/wealth_data_for_import2.csv') as f2:
+        with open('../data/wealth_data_for_import2.csv') as f2:
             self.data = pd.read_csv(f2, encoding = 'unicode_escape')        
         y = model_params["start_year"]
         self.idx_begin = min((self.data[self.data["year"]==y].index.values))
