+++ src/plot_FIG1.py
@@ -18,7 +18,7 @@ import pandas as pd
 
 #%%
 
-with open('./data/wealth_data_for_import.csv') as f:
+with open('../data/wealth_data_for_import.csv') as f:
     d1 = pd.read_csv(f, encoding = 'unicode_escape')  
     
 colors = ["tab:red", "tab:blue", "grey", "tab:olive"]
